==> Notebooks/1_ComparableCompanies.ipynb <==
### Lybraries:

- Pymongo (MongoClient)
- Pandas

%pip install pymongo


from pymongo import MongoClient
import pandas as pd

### Exploring collection: 'companies' 

Calling the collection from MongoCompass

client = MongoClient("localhost:27017")

db = client.get_database("ironhack")

c = db.get_collection("companies")

Searching for similar companies. Ideally: gaming, less than 100 employees, startups (less than five years since founding, collection was updated in 2013), web...
This way we will select the three that are more similar and explore their cities to match the conditions.

#We are going to stablish several conditions to find comparable companies, and realize querys to find the most similar ones
cond = {'founded_year':{'$gt':2008}}
cond1 = {'number_of_employees':{'$lte':100}}
cond2 ={'category_code': 'games_video'}
cond6 = {'number_of_employees':{'$gte':30}}

#query = {'$and':[{'$or':[cond2,cond3,cond4,cond5]},cond, cond1,cond6]}
query = {'$and':[cond2,cond, cond1,cond6]}

proy = {'_id':0,'name':1, 'number_of_employees':1, 'founded_year':1, 'category_code' : 1, 'offices.city':1}

companies=list(c.find(query,proy).sort('offices.0.city',-1))

Lot of the companies found dont have the city information, so we are going to get the ones that have it

comparables = []
for i in companies:
    if len(i) > 4:
        comparables.append(i)

sim_comp = pd.DataFrame(comparables)


sim_comp

### Our three cities are
    - Vancouver for Tiny Speck
    - Beverly Hills for Social Gaming Network
    - Irvine  for OneScreen

sim_comp['latitude'] = [49.276345, 34.069722, 33.669445]

sim_comp['longitude'] = [-123.117337, -118.405556, -117.823059]

sim_comp.to_csv('../data/similarcompanies.csv')
