==> Notebooks/4_Maps.ipynb <==
import pandas as pd
import geopandas as gpd
from keplergl import KeplerGl

bh = pd.read_csv('../data/BeverlyHillsCompleteData.csv')

bh.sample(5)

gdf = gpd.GeoDataFrame(bh, geometry=gpd.points_from_xy(bh.longitud, bh.latitud))

gdf

map_1= KeplerGl(height=700, weight = 500, data={'BeverlyHills': gdf})

map_1

#
